requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
google-cloud-texttospeech==2.14.1
openai==0.28.0
